Fixing bug.
git-svn-id: https://code.ua.pt/svn/security2013-p1g1@53 a478ab85-490a-429d-afbb-08a0a9d09df7

+++ proj1/Flexible Cryptographic File Protection/src/fcfp/util/cast/ByteCast.java
@@ -33,7 +33,7 @@ public class ByteCast {
      */
     public static long byteArray2Long(byte[] data) {
         long num = 0;
-        for (int i = 0; i < Long.SIZE; i++) {
+        for (int i = 0; i < Byte.SIZE; i++) {
             num += (data[i] << (i << 3));
         }
         return num;
